CREATING INITIAL CONTROLLER AND CLASSES
QuizzController is created, at the moment, with one main method (createQuestion) which is accessible through /quizz/question. This method is not able to create a question for now, but it can already process the JSON received from TMDB API to make an Object out of it (a movie for instance).