CREATING/UPDATING DESERIALIZER
To be able to use the Json from TMDB API, we have to read its content and create object out of it. It means creating a Deserializer for every type of object expected in return of TMDB API call.
In this case, we create a deserializer for an Actor and we modify the Movie one to add a method used in QuizzController to call deserialization